// src/domain/models/artist.ts
import { Image } from "./image";

export interface Artist {
  external_urls: {
    spotify: string;
  };
  followers: {
    href?: string;
    total: number;
  };
  genres: string[];
  href: string;
  id: string;
  images: Image[];
  name: string;
  popularity: number;
  type: string;
  uri: string;
}